Adding an Appointments screen to the app (#26)
# *Name of the PR*
Adding an Appointments screen to the app

## :recycle: Current situation & Problem
*Link any open issues or pull requests (PRs) related to this PR. Please
ensure that all non-trivial PRs are first tracked and discussed in an
existing GitHub issue or discussion.*
There was no screen to display the appointments on the app, and as a
team, we found it to be anon-negotiable for our app's objectives. As a
result, I have implemented a template tab that shares appointment
details with the user. This PR is all front-end based and will be
updated as the backend changes to store the necessary information on the
appointments.

## :gear: Release Notes 
*Add a bullet point list summary of the feature and possible migration
guides if this is a breaking change so this section can be added to the
release notes.*
*Include code snippets that provide examples of the feature implemented
or links to the documentation if it appends or changes the public
interface.*
- home screen is now the new Appointments screen
- template screen holds two boxes for the two upcoming appointments for
users
- the appointment contains date, time, and required info details as well
as a dummy button that will eventually allow you to edit appointment
details
- there is also a map view at the bottom that provides the address and
an Apple Maps view of the hospital (can add details about how to get
there later)


## :books: Documentation
*Please ensure that you properly document any additions in conformance
to [Spezi Documentation
Guide](https://github.com/StanfordSpezi/.github/blob/main/DOCUMENTATIONGUIDE.md).*
*You can use this section to describe your solution, but we encourage
contributors to document your reasoning and changes using in-line
documentation.*
At the moment, since these changes have not been reviewed with the team,
I don't think documentation is necessary, but I will update this section
as we go forward after a preliminary review.


## :white_check_mark: Testing
*Please ensure that the PR meets the testing requirements set by CodeCov
and that new functionality is appropriately tested.*
*This section describes important information about the tests and why
some elements might not be testable.*
There has been no testing done yet, but the implementations are all
design and front-end based so testing is not quite as relevant as there
is no data involved.

## :pencil: Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/CS342/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/CS342/.github/blob/main/CONTRIBUTING.md):
- [ ] I agree to follow the [Code of
Conduct](https://github.com/CS342/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/CS342/.github/blob/main/CONTRIBUTING.md).

---------

Co-authored-by: abhat01 <akanshya@slac.stanford.edu>